Configure jekyll directory locations
Changed the source dir to www (it's parent directory) and adjust all of the other
paths to go back down into the _jekyll directory from www.

Also exclude the _jekyll directory itself from output.

This is to keep the content outside of the jekyll root, so actual content
is not mixed in with jekyll stuffs.